requirements.txt
amqp==5.3.1 asgiref==3.10.0 attrs==25.4.0 billiard==4.2.2 black==25.9.0 celery==5.5.3 certifi==2025.10.5 cfgv==3.4.0 click==8.3.0 click-didyoumean==0.3.1 click-plugins==1.1.1.2
click-repl==0.3.0 colorama==0.4.6 distlib==0.4.0 Django==5.2.7 django-cors-headers==4.9.0 django-debug-toolbar==6.0.0 django-extensions==4.1 django-filter==25.2 djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1 dotenv==0.9.9 drf-spectacular==0.28.0 factory_boy==3.3.3 Faker==37.12.0 filelock==3.20.0 flake8==7.3.0 gunicorn==23.0.0 identify==2.6.15
inflection==0.5.1 iniconfig==2.3.0 isort==7.0.0 jsonschema==4.25.1 jsonschema-specifications==2025.9.1 kombu==5.5.4 mccabe==0.7.0 mypy_extensions==1.1.0 nodeenv==1.9.1
packaging==25.0 pathspec==0.12.1 pillow==12.0.0 platformdirs==4.5.0 pluggy==1.6.0 pre_commit==4.3.0 prompt_toolkit==3.0.52 psycopg2-binary==2.9.11 pycodestyle==2.14.0 pyflakes==3.4.0
Pygments==2.19.2 PyJWT==2.10.1 pytest==8.4.2 pytest-django==4.11.1 python-dateutil==2.9.0.post0 python-dotenv==1.1.1 pytokens==0.2.0 PyYAML==6.0.3 redis==7.0.0 referencing==0.37.0
rpds-py==0.28.0 sentry-sdk==2.42.1 six==1.17.0 sqlparse==0.5.3 typing_extensions==4.15.0 tzdata==2025.2 uritemplate==4.2.0 urllib3==2.5.0 vine==5.1.0 virtualenv==20.35.3 wcwidth==0.2.14